Move volume visitor to a separate file
