+++ src/celeritas/CMakeLists.txt
@@ -103,10 +103,11 @@ if(CELERITAS_USE_Geant4)
     ext/detail/GeantExceptionHandler.cc
     ext/detail/GeantGeoExporter.cc
     ext/detail/GeantLoggerAdapter.cc
-    ext/detail/GeantPhysicsList.cc
-    ext/detail/GeantModelImporter.cc
-    ext/detail/GeantProcessImporter.cc
     ext/detail/GeantMicroXsCalculator.cc
+    ext/detail/GeantModelImporter.cc
+    ext/detail/GeantPhysicsList.cc
+    ext/detail/GeantProcessImporter.cc
+    ext/detail/GeantVolumeVisitor.cc
   )
   target_link_libraries(celeritas_geant4
     PRIVATE Celeritas::corecel XercesC::XercesC ${Geant4_LIBRARIES}
